+++ pajbot/bot.py
@@ -32,7 +32,7 @@ from pajbot.managers.irc import IRCManager
 from pajbot.managers.kvi import KVIManager
 from pajbot.managers.redis import RedisManager
 from pajbot.managers.schedule import ScheduleManager
-from pajbot.managers.twitter import TwitterManager
+from pajbot.managers.twitter import TwitterManager, PBTwitterManager
 from pajbot.managers.user_ranks_refresh import UserRanksRefreshManager
 from pajbot.managers.websocket import WebSocketManager
 from pajbot.migration.db import DatabaseMigratable
@@ -196,7 +196,10 @@ class Bot:
         self.emote_manager = EmoteManager(self.twitch_v5_api, self.action_queue)
         self.epm_manager = EpmManager()
         self.ecount_manager = EcountManager()
-        self.twitter_manager = TwitterManager(self)
+        if "twitter" in self.config and self.config["twitter"].get("streaming_type", "twitter") == "tweet-provider":
+            self.twitter_manager = PBTwitterManager(self)
+        else:
+            self.twitter_manager = TwitterManager(self)
         self.module_manager = ModuleManager(self.socket_manager, bot=self).load()
         self.commands = CommandManager(
             socket_manager=self.socket_manager, module_manager=self.module_manager, bot=self