+++ pajbot/utils/__init__.py
@@ -14,7 +14,7 @@ from .parse_points_amount import parse_points_amount
 from .print_traceback import print_traceback
 from .remove_none_values import remove_none_values
 from .split_into_chunks_with_prefix import split_into_chunks_with_prefix
-from .stringify_tweet import stringify_tweet
+from .stringify_tweet import stringify_tweet, tweet_provider_stringify_tweet
 from .time_ago import time_ago
 from .time_limit import time_limit
 from .time_method import time_method